Small refactoring of Auth API
