+++ auth-api/.gitignore
@@ -1,1 +1,2 @@
 vendor
+auth-api
